Move user components to /home
It decouples user components from the app nav, which must not be part of the main element.
