+++ src/components/home/user-content/cv-section/cv-section.scss
@@ -2,8 +2,8 @@
 /* SPDX-License-Identifier: MIT */
 /* This file is part of https://github.com/tobiasbriones/my-page. */
 
-@import "../../../../../styles/h3";
-@import "../../../../../styles/h4";
+@import "../../../styles/h3";
+@import "../../../styles/h4";
 
 :host {
   display: block;
