+++ src/components/home/user-content/user-engineering/user-engineering.scss
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: MIT
 // This file is part of https://github.com/tobiasbriones/my-page.
 
-@import "../../../../../styles/h2";
+@import "../../../styles/h2";
 
 :host {
   display: block;
