+++ src/components/home/user-content/user-engineering/user-engineering.tsx
@@ -1,6 +1,6 @@
 import { Component, h, Host, Prop } from '@stencil/core';
-import { emptyEngineering, Engineering } from '../../../../../../user';
-import { parseMarkdown } from '../../../../../../markdown/markdown';
+import { emptyEngineering, Engineering } from '../../../../user';
+import { parseMarkdown } from '../../../../markdown/markdown';
 
 @Component({
   tag: 'me-user-engineering',
